--- a/pyproject.toml
+++ b/pyproject.toml
@@ -24,6 +24,7 @@ ruff = "^0.3.3"
 snakeviz = "^2.2.0"
 astropy = "^6.0.1"
 setuptools = "^70.0.0"
+viztracer = "^0.16.3"
 
 
 [build-system]