--- a/src/prx/main.py
+++ b/src/prx/main.py
@@ -18,10 +18,10 @@ log = helpers.get_logger(__name__)
 
 @helpers.timeit
 def write_prx_file(
-        prx_header: dict,
-        prx_records: pd.DataFrame,
-        file_name_without_extension: Path,
-        output_format: str,
+    prx_header: dict,
+    prx_records: pd.DataFrame,
+    file_name_without_extension: Path,
+    output_format: str,
 ):
     output_writers = {"jsonseq": write_json_text_sequence_file, "csv": write_csv_file}
     if output_format not in output_writers.keys():
@@ -30,7 +30,7 @@ def write_prx_file(
 
 
 def write_json_text_sequence_file(
-        prx_header: dict, prx_records: pd.DataFrame, file_name_without_extension: Path
+    prx_header: dict, prx_records: pd.DataFrame, file_name_without_extension: Path
 ):
     output_file = Path(
         f"{str(file_name_without_extension)}.{constants.cPrxJsonTextSequenceFileExtension}"
@@ -46,7 +46,7 @@ def write_json_text_sequence_file(
             epoch = pd.Timestamp(epoch)
             epoch_obs = prx_records[
                 prx_records["time_of_reception_in_receiver_time"] == epoch
-                ]
+            ]
             record = {
                 "time_of_reception_in_receiver_time": epoch.strftime(
                     "%Y:%m:%dT%H:%M:%S.%f"
@@ -73,7 +73,7 @@ def write_json_text_sequence_file(
 
 
 def write_csv_file(
-        prx_header: dict, flat_records: pd.DataFrame, file_name_without_extension: Path
+    prx_header: dict, flat_records: pd.DataFrame, file_name_without_extension: Path
 ):
     output_file = Path(
         f"{str(file_name_without_extension)}.{constants.cPrxCsvFileExtension}"
@@ -182,23 +182,23 @@ def build_metadata(input_files):
 
 
 def check_assumptions(
-        rinex_3_obs_file,
+    rinex_3_obs_file,
 ):
     obs_header = georinex.rinexheader(rinex_3_obs_file)
     if "RCV CLOCK OFFS APPL" in obs_header.keys():
         assert (
-                obs_header["RCV CLOCK OFFS APPL"].strip() == "0"
+            obs_header["RCV CLOCK OFFS APPL"].strip() == "0"
         ), "Handling of 'RCV CLOCK OFFS APPL' != 0 not implemented yet."
     assert (
-            obs_header["TIME OF FIRST OBS"].split()[-1].strip() == "GPS"
+        obs_header["TIME OF FIRST OBS"].split()[-1].strip() == "GPS"
     ), "Handling of observation files using time scales other than GPST not implemented yet."
 
 
 @helpers.timeit
 def build_records(
-        rinex_3_obs_file,
-        rinex_3_ephemerides_files,
-        approximate_receiver_ecef_position_m,
+    rinex_3_obs_file,
+    rinex_3_ephemerides_files,
+    approximate_receiver_ecef_position_m,
 ):
     approximate_receiver_ecef_position_m = np.array(
         approximate_receiver_ecef_position_m
@@ -304,14 +304,14 @@ def build_records(
         doy = int(file.name[16:19])
         day_query = query.loc[
             (
-                    query.query_time_isagpst
-                    >= pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy - 1)
+                query.query_time_isagpst
+                >= pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy - 1)
             )
             & (
-                    query.query_time_isagpst
-                    < pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy)
+                query.query_time_isagpst
+                < pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy)
             )
-            ]
+        ]
         if day_query.empty:
             continue
         log.info(f"Computing satellite states for {year}-{doy:03d}")
@@ -416,7 +416,7 @@ def build_records(
             return np.nan
         return constants.carrier_frequencies_hz()[row.satellite[0]][
             "L" + row.observation_type[1]
-            ][row["frequency_slot"]]
+        ][row["frequency_slot"]]
 
     flat_obs.loc[:, "carrier_frequency_hz"] = flat_obs.apply(
         signal_2_carrier_frequency, axis=1
@@ -436,12 +436,12 @@ def build_records(
 
         # Selection criteria: time of emission belonging to the day of the current NAV file
         mask = (
-                       flat_obs.time_of_emission_isagpst
-                       >= pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy - 1)
-               ) & (
-                       flat_obs.time_of_emission_isagpst
-                       < pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy)
-               )
+            flat_obs.time_of_emission_isagpst
+            >= pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy - 1)
+        ) & (
+            flat_obs.time_of_emission_isagpst
+            < pd.Timestamp(year=year, month=1, day=1) + pd.Timedelta(days=doy)
+        )
         if "IONOSPHERIC CORR" in nav_header_dict[f"{year:03d}" + f"{doy:03d}"]:
             log.info(f"Computing iono delay for {year}-{doy:03d}")
             flat_obs.loc[
@@ -460,9 +460,9 @@ def build_records(
                 latitude_user_rad,
                 longitude_user_rad,
             ) * (
-                        constants.carrier_frequencies_hz()["G"]["L1"][1] ** 2
-                        / flat_obs.loc[mask].carrier_frequency_hz ** 2
-                )
+                constants.carrier_frequencies_hz()["G"]["L1"][1] ** 2
+                / flat_obs.loc[mask].carrier_frequency_hz ** 2
+            )
         else:
             logging.warning(f"Missing iono model parameters for day {doy:03d}")
             flat_obs.loc[
@@ -507,7 +507,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         prog="prx",
         description="prx processes RINEX observations, computes a few useful things such as satellite position, "
-                    "relativistic effects etc. and outputs everything to a text file in a convenient format.",
+        "relativistic effects etc. and outputs everything to a text file in a convenient format.",
         epilog="P.S. GNSS rules!",
     )
     parser.add_argument(