This PR

- Improves the number of disk cache hits by accessing the cache independently of file name, i.e. only considering the file content hash
- Removes the timestamp left by gfzrnx when repairing a file, which circumvents caching
- Vectorizes matching ephemerides to observations. Before, the `select_ephemerides` function was taking the most time, apart from the georinex parsing code, about 1.5' for a 15' file. With this change, that went down to 1.9 seconds.
- Removes the PRX_NO_CACHING environment variable and the associated code, which made things slower than they need to be by repeatedly flushing the cache when using multiprocessing. Manually deleting the disk cache folder has the same effect.
- Parses nav and obs files in parallel

# Testing

Running on a 15-minute observation file

After deleting the disk cache folder
```
python3 src/prx/main.py --observation_file_path   96.16s user 18.42s system 173% cpu 1:05.93 total
```
-> processing speed about 15*60/66 = 13.6

With caching
```
python3 src/prx/main.py --observation_file_path   28.91s user 8.79s system 142% cpu 26.530 total
```
-> processing speed about 15*60/27 = 33

Without caching on `main` I get a crash here https://github.com/jtec/prx/blob/9ccd386d6208b8ac1222df1d0acbaff1f15b3bbf/src/prx/rinex_nav/evaluate.py#L660 but until then it takes about 2'17'' (processing speed about 15*60/137 = 6.5):

```
PRX_NO_CACHING=true python3 src/prx/main.py --observation_file_path   251.36s user 54.30s system 222% cpu 2:17.67 total
```