@plutonheaven 2024-06-25
I added some of the changes I did in another branch, to improve the process. I believe it is coherent with the speedup goal of this branch.
If you prefer to decorrelate those changes, you may revert to cc81ad04eaeab47ee98915124675879122260d8b.